Allow you to specify a context when subscribing to a serie
This allows the UI to specify an organisational unit as context when
subscribing to a serie. We need the UI to tell us what the context is
as series can be borrowed under multiple organisational units and the
backend has no way of knowing which one is the relevant one.